+++ src/model/index.js
@@ -1,0 +1,1 @@
+export { default as MessageModel } from "./MessageModel"