+++ src/scenes/manage/searchresult/SearchResult.js
@@ -47,8 +47,8 @@ export default class SearchResult extends Component {
 			self.setState({
 				hasSeries: true,
 				processing: false
-			});
-		});
+			})
+		})
 	}
 
 	removeSeries() {
@@ -72,10 +72,10 @@ export default class SearchResult extends Component {
 			return 'bright-squares.png';
 		}
 	}
-	
+
 	render() {
 		const series = this.props.series;
-		
+
 		const actions = () => {
 			if (this.state.processing) {
 				return;
@@ -83,27 +83,27 @@ export default class SearchResult extends Component {
 			if (this.state.hasSeries) {
 				return (
 					<div>
-						<button onClick={ this.addSeries }><span className="fa fa-refresh"></span></button>
-						<button onClick={ this.removeSeries }><span className="fa fa-trash"></span></button>
-						<Link className="view-link" to={ `/view/${series.id}` }><span className="fa fa-tv"></span></Link>
+						<button onClick={this.addSeries}><span className="fa fa-refresh"></span></button>
+						<button onClick={this.removeSeries}><span className="fa fa-trash"></span></button>
+						<Link className="view-link" to={`/view/${series.id}`}><span className="fa fa-tv"></span></Link>
 					</div>
 				);
 			}
-			return (<button onClick={ this.addSeries }><span className="fa fa-plus"></span></button>);
+			return (<button onClick={this.addSeries}><span className="fa fa-plus"></span></button>);
 		}
 
 		return (
 			<div className="series-result">
-				<img src={ this.getImageSrc(series) } alt="" />
+				<img src={this.getImageSrc(series)} alt="" />
 				<div className="series-title-wrapper">
-					<div className="series-title">{ series.name }</div>
+					<div className="series-title">{series.name}</div>
 				</div>
-				<div className="airing">{ moment(series.first_air_date).format('DD.MM.YYYY') }</div>
+				<div className="airing">{moment(series.first_air_date).format('DD.MM.YYYY')}</div>
 				<div className="actions">
-					{ actions() }
+					{actions()}
 				</div>
-				<div className="series-description">{ series.overview }</div>
-	  		</div>
+				<div className="series-description">{series.overview}</div>
+			</div>
 		)
 	}
 }